webapi,trcapi,make: Add support for auth-less builds
